+++ action_server/tests/action_server_tests/resources/no_conda/calculator/calculator_tasks.py
@@ -30,3 +30,10 @@ def neverending_action() -> str:
 
     time.sleep(500)
     raise RuntimeError("Should never get here...")
+
+
+@action
+def bad_return_none():
+    """
+    Return None when string is expected
+    """